Refactoring controller codes
