+++ controllers/templateinstance/templateinstance_controller_test.go
@@ -1,4 +1,4 @@
-package controllers
+package templateinstance
 
 import (
 	"context"
