+++ internal/type.go
@@ -1,0 +1,6 @@
+package internal
+
+const (
+	ClaimFinalizer = "clustertemplateclaims.tmax.io/finalizer"
+	ClaimLabel     = "clustertemplateclaims.tmax.io/claim"
+)
