+++ main.go
@@ -18,6 +18,10 @@ package main
 
 import (
 	"flag"
+	"github.com/tmax-cloud/template-operator/controllers/clustertemplate"
+	"github.com/tmax-cloud/template-operator/controllers/clustertemplateclaim"
+	"github.com/tmax-cloud/template-operator/controllers/template"
+	"github.com/tmax-cloud/template-operator/controllers/templateinstance"
 	"os"
 
 	"k8s.io/apimachinery/pkg/runtime"
@@ -28,7 +32,6 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/log/zap"
 
 	tmaxiov1 "github.com/tmax-cloud/template-operator/api/v1"
-	"github.com/tmax-cloud/template-operator/controllers"
 	// +kubebuilder:scaffold:imports
 )
 
@@ -72,7 +75,7 @@ func main() {
 		os.Exit(1)
 	}
 
-	if err = (&controllers.TemplateReconciler{
+	if err = (&template.TemplateReconciler{
 		Client: mgr.GetClient(),
 		Log:    ctrl.Log.WithName("controllers").WithName("Template"),
 		Scheme: mgr.GetScheme(),
@@ -80,7 +83,7 @@ func main() {
 		setupLog.Error(err, "unable to create controller", "controller", "Template")
 		os.Exit(1)
 	}
-	if err = (&controllers.ClusterTemplateReconciler{
+	if err = (&clustertemplate.ClusterTemplateReconciler{
 		Client: mgr.GetClient(),
 		Log:    ctrl.Log.WithName("controllers").WithName("ClusterTemplate"),
 		Scheme: mgr.GetScheme(),
@@ -88,7 +91,7 @@ func main() {
 		setupLog.Error(err, "unable to create controller", "controller", "ClusterTemplate")
 		os.Exit(1)
 	}
-	if err = (&controllers.TemplateInstanceReconciler{
+	if err = (&templateinstance.TemplateInstanceReconciler{
 		Client: mgr.GetClient(),
 		Log:    ctrl.Log.WithName("controllers").WithName("TemplateInstance"),
 		Scheme: mgr.GetScheme(),
@@ -96,7 +99,7 @@ func main() {
 		setupLog.Error(err, "unable to create controller", "controller", "TemplateInstance")
 		os.Exit(1)
 	}
-	if err = (&controllers.ClusterTemplateClaimReconciler{
+	if err = (&clustertemplateclaim.ClusterTemplateClaimReconciler{
 		Client: mgr.GetClient(),
 		Log:    ctrl.Log.WithName("controllers").WithName("ClusterTemplateClaim"),
 		Scheme: mgr.GetScheme(),
